[hotfix][python] Remove unless classes for the old planner
