+++ flink-python/src/main/java/org/apache/flink/table/runtime/arrow/ArrowReader.java
@@ -19,15 +19,38 @@
 package org.apache.flink.table.runtime.arrow;
 
 import org.apache.flink.annotation.Internal;
+import org.apache.flink.table.data.ColumnarRowData;
+import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.data.vector.ColumnVector;
+import org.apache.flink.table.data.vector.VectorizedColumnBatch;
+import org.apache.flink.util.Preconditions;
 
-/**
- * Reader which deserialize the Arrow format data to the Flink rows.
- *
- * @param <OUT> Type of the deserialized row.
- */
+/** {@link ArrowReader} which read the underlying Arrow format data as {@link RowData}. */
 @Internal
-public interface ArrowReader<OUT> {
+public final class ArrowReader {
+
+    /**
+     * An array of vectors which are responsible for the deserialization of each column of the rows.
+     */
+    private final ColumnVector[] columnVectors;
+
+    /** Reusable row used to hold the deserialized result. */
+    private ColumnarRowData reuseRow;
+
+    public ArrowReader(ColumnVector[] columnVectors) {
+        this.columnVectors = Preconditions.checkNotNull(columnVectors);
+        this.reuseRow = new ColumnarRowData();
+    }
+
+    /** Gets the column vectors. */
+    public ColumnVector[] getColumnVectors() {
+        return columnVectors;
+    }
 
     /** Read the specified row from underlying Arrow format data. */
-    OUT read(int rowId);
+    public RowData read(int rowId) {
+        reuseRow.setVectorizedColumnBatch(new VectorizedColumnBatch(columnVectors));
+        reuseRow.setRowId(rowId);
+        return reuseRow;
+    }
 }
