--- flink-python/src/main/java/org/apache/flink/table/runtime/arrow/writers/RowRowWriter.java
@@ -1,69 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.table.runtime.arrow.writers;
-
-import org.apache.flink.annotation.Internal;
-import org.apache.flink.types.Row;
-
-import org.apache.arrow.vector.complex.StructVector;
-
-/** {@link ArrowFieldWriter} for Row. */
-@Internal
-public final class RowRowWriter extends ArrowFieldWriter<Row> {
-
-    private final ArrowFieldWriter<Row>[] fieldsWriters;
-    private final Row nullRow;
-
-    public RowRowWriter(StructVector structVector, ArrowFieldWriter<Row>[] fieldsWriters) {
-        super(structVector);
-        this.fieldsWriters = fieldsWriters;
-        this.nullRow = new Row(fieldsWriters.length);
-    }
-
-    @Override
-    public void doWrite(Row value, int ordinal) {
-        Row row;
-        if (value.getField(ordinal) == null) {
-            row = nullRow;
-            ((StructVector) getValueVector()).setNull(getCount());
-        } else {
-            row = (Row) value.getField(ordinal);
-            ((StructVector) getValueVector()).setIndexDefined(getCount());
-        }
-        for (int i = 0; i < fieldsWriters.length; i++) {
-            fieldsWriters[i].write(row, i);
-        }
-    }
-
-    @Override
-    public void finish() {
-        super.finish();
-        for (ArrowFieldWriter<?> fieldsWriter : fieldsWriters) {
-            fieldsWriter.finish();
-        }
-    }
-
-    @Override
-    public void reset() {
-        super.reset();
-        for (ArrowFieldWriter<?> fieldsWriter : fieldsWriters) {
-            fieldsWriter.reset();
-        }
-    }
-}
