+++ flink-python/src/test/java/org/apache/flink/table/runtime/arrow/ArrowUtilsTest.java
@@ -18,7 +18,7 @@
 
 package org.apache.flink.table.runtime.arrow;
 
-import org.apache.flink.api.java.tuple.Tuple6;
+import org.apache.flink.api.java.tuple.Tuple5;
 import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.data.vector.ColumnVector;
@@ -37,7 +37,6 @@ import org.apache.flink.table.runtime.arrow.vectors.ArrowTimestampColumnVector;
 import org.apache.flink.table.runtime.arrow.vectors.ArrowTinyIntColumnVector;
 import org.apache.flink.table.runtime.arrow.vectors.ArrowVarBinaryColumnVector;
 import org.apache.flink.table.runtime.arrow.vectors.ArrowVarCharColumnVector;
-import org.apache.flink.table.runtime.arrow.vectors.RowDataArrowReader;
 import org.apache.flink.table.runtime.arrow.writers.ArrayWriter;
 import org.apache.flink.table.runtime.arrow.writers.ArrowFieldWriter;
 import org.apache.flink.table.runtime.arrow.writers.BigIntWriter;
@@ -47,21 +46,6 @@ import org.apache.flink.table.runtime.arrow.writers.DecimalWriter;
 import org.apache.flink.table.runtime.arrow.writers.DoubleWriter;
 import org.apache.flink.table.runtime.arrow.writers.FloatWriter;
 import org.apache.flink.table.runtime.arrow.writers.IntWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowArrayWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowBigIntWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowBooleanWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowDateWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowDecimalWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowDoubleWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowFloatWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowIntWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowRowWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowSmallIntWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowTimeWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowTimestampWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowTinyIntWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowVarBinaryWriter;
-import org.apache.flink.table.runtime.arrow.writers.RowVarCharWriter;
 import org.apache.flink.table.runtime.arrow.writers.RowWriter;
 import org.apache.flink.table.runtime.arrow.writers.SmallIntWriter;
 import org.apache.flink.table.runtime.arrow.writers.TimeWriter;
@@ -114,8 +98,7 @@ import static org.junit.Assert.assertEquals;
 /** Tests for {@link ArrowUtils}. */
 public class ArrowUtilsTest {
 
-    private static List<Tuple6<String, LogicalType, ArrowType, Class<?>, Class<?>, Class<?>>>
-            testFields;
+    private static List<Tuple5<String, LogicalType, ArrowType, Class<?>, Class<?>>> testFields;
     private static RowType rowType;
     private static BufferAllocator allocator;
 
@@ -123,218 +106,194 @@ public class ArrowUtilsTest {
     public static void init() {
         testFields = new ArrayList<>();
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f1",
                         new TinyIntType(),
                         new ArrowType.Int(8, true),
-                        RowTinyIntWriter.class,
                         TinyIntWriter.TinyIntWriterForRow.class,
                         ArrowTinyIntColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f2",
                         new SmallIntType(),
                         new ArrowType.Int(8 * 2, true),
-                        RowSmallIntWriter.class,
                         SmallIntWriter.SmallIntWriterForRow.class,
                         ArrowSmallIntColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f3",
                         new IntType(),
                         new ArrowType.Int(8 * 4, true),
-                        RowIntWriter.class,
                         IntWriter.IntWriterForRow.class,
                         ArrowIntColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f4",
                         new BigIntType(),
                         new ArrowType.Int(8 * 8, true),
-                        RowBigIntWriter.class,
                         BigIntWriter.BigIntWriterForRow.class,
                         ArrowBigIntColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f5",
                         new BooleanType(),
                         new ArrowType.Bool(),
-                        RowBooleanWriter.class,
                         BooleanWriter.BooleanWriterForRow.class,
                         ArrowBooleanColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f6",
                         new FloatType(),
                         new ArrowType.FloatingPoint(FloatingPointPrecision.SINGLE),
-                        RowFloatWriter.class,
                         FloatWriter.FloatWriterForRow.class,
                         ArrowFloatColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f7",
                         new DoubleType(),
                         new ArrowType.FloatingPoint(FloatingPointPrecision.DOUBLE),
-                        RowDoubleWriter.class,
                         DoubleWriter.DoubleWriterForRow.class,
                         ArrowDoubleColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f8",
                         new VarCharType(),
                         ArrowType.Utf8.INSTANCE,
-                        RowVarCharWriter.class,
                         VarCharWriter.VarCharWriterForRow.class,
                         ArrowVarCharColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f9",
                         new VarBinaryType(),
                         ArrowType.Binary.INSTANCE,
-                        RowVarBinaryWriter.class,
                         VarBinaryWriter.VarBinaryWriterForRow.class,
                         ArrowVarBinaryColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f10",
                         new DecimalType(10, 3),
                         new ArrowType.Decimal(10, 3),
-                        RowDecimalWriter.class,
                         DecimalWriter.DecimalWriterForRow.class,
                         ArrowDecimalColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f11",
                         new DateType(),
                         new ArrowType.Date(DateUnit.DAY),
-                        RowDateWriter.class,
                         DateWriter.DateWriterForRow.class,
                         ArrowDateColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f13",
                         new TimeType(0),
                         new ArrowType.Time(TimeUnit.SECOND, 32),
-                        RowTimeWriter.class,
                         TimeWriter.TimeWriterForRow.class,
                         ArrowTimeColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f14",
                         new TimeType(2),
                         new ArrowType.Time(TimeUnit.MILLISECOND, 32),
-                        RowTimeWriter.class,
                         TimeWriter.TimeWriterForRow.class,
                         ArrowTimeColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f15",
                         new TimeType(4),
                         new ArrowType.Time(TimeUnit.MICROSECOND, 64),
-                        RowTimeWriter.class,
                         TimeWriter.TimeWriterForRow.class,
                         ArrowTimeColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f16",
                         new TimeType(8),
                         new ArrowType.Time(TimeUnit.NANOSECOND, 64),
-                        RowTimeWriter.class,
                         TimeWriter.TimeWriterForRow.class,
                         ArrowTimeColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f17",
                         new LocalZonedTimestampType(0),
                         new ArrowType.Timestamp(TimeUnit.SECOND, null),
-                        RowTimestampWriter.class,
                         TimestampWriter.TimestampWriterForRow.class,
                         ArrowTimestampColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f18",
                         new LocalZonedTimestampType(2),
                         new ArrowType.Timestamp(TimeUnit.MILLISECOND, null),
-                        RowTimestampWriter.class,
                         TimestampWriter.TimestampWriterForRow.class,
                         ArrowTimestampColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f19",
                         new LocalZonedTimestampType(4),
                         new ArrowType.Timestamp(TimeUnit.MICROSECOND, null),
-                        RowTimestampWriter.class,
                         TimestampWriter.TimestampWriterForRow.class,
                         ArrowTimestampColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f20",
                         new LocalZonedTimestampType(8),
                         new ArrowType.Timestamp(TimeUnit.NANOSECOND, null),
-                        RowTimestampWriter.class,
                         TimestampWriter.TimestampWriterForRow.class,
                         ArrowTimestampColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f21",
                         new TimestampType(0),
                         new ArrowType.Timestamp(TimeUnit.SECOND, null),
-                        RowTimestampWriter.class,
                         TimestampWriter.TimestampWriterForRow.class,
                         ArrowTimestampColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f22",
                         new TimestampType(2),
                         new ArrowType.Timestamp(TimeUnit.MILLISECOND, null),
-                        RowTimestampWriter.class,
                         TimestampWriter.TimestampWriterForRow.class,
                         ArrowTimestampColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f23",
                         new TimestampType(4),
                         new ArrowType.Timestamp(TimeUnit.MICROSECOND, null),
-                        RowTimestampWriter.class,
                         TimestampWriter.TimestampWriterForRow.class,
                         ArrowTimestampColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f24",
                         new TimestampType(8),
                         new ArrowType.Timestamp(TimeUnit.NANOSECOND, null),
-                        RowTimestampWriter.class,
                         TimestampWriter.TimestampWriterForRow.class,
                         ArrowTimestampColumnVector.class));
 
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f25",
                         new ArrayType(new VarCharType()),
                         ArrowType.List.INSTANCE,
-                        RowArrayWriter.class,
                         ArrayWriter.ArrayWriterForRow.class,
                         ArrowArrayColumnVector.class));
 
@@ -353,17 +312,15 @@ public class ArrowUtilsTest {
                                                         new RowType.RowField(
                                                                 "e2", new VarCharType())))))));
         testFields.add(
-                Tuple6.of(
+                Tuple5.of(
                         "f26",
                         rowFieldType,
                         ArrowType.Struct.INSTANCE,
-                        RowRowWriter.class,
                         RowWriter.RowWriterForRow.class,
                         ArrowRowColumnVector.class));
 
         List<RowType.RowField> rowFields = new ArrayList<>();
-        for (Tuple6<String, LogicalType, ArrowType, Class<?>, Class<?>, Class<?>> field :
-                testFields) {
+        for (Tuple5<String, LogicalType, ArrowType, Class<?>, Class<?>> field : testFields) {
             rowFields.add(new RowType.RowField(field.f0, field.f1));
         }
         rowType = new RowType(rowFields);
@@ -385,24 +342,24 @@ public class ArrowUtilsTest {
     }
 
     @Test
-    public void testCreateRowDataArrowReader() {
+    public void testCreateArrowReader() {
         VectorSchemaRoot root =
                 VectorSchemaRoot.create(ArrowUtils.toArrowSchema(rowType), allocator);
-        RowDataArrowReader reader = ArrowUtils.createRowDataArrowReader(root, rowType);
+        ArrowReader reader = ArrowUtils.createArrowReader(root, rowType);
         ColumnVector[] columnVectors = reader.getColumnVectors();
         for (int i = 0; i < columnVectors.length; i++) {
-            assertEquals(testFields.get(i).f5, columnVectors[i].getClass());
+            assertEquals(testFields.get(i).f4, columnVectors[i].getClass());
         }
     }
 
     @Test
-    public void testCreateRowDataArrowWriter() {
+    public void testCreateArrowWriter() {
         VectorSchemaRoot root =
                 VectorSchemaRoot.create(ArrowUtils.toArrowSchema(rowType), allocator);
         ArrowWriter<RowData> writer = ArrowUtils.createRowDataArrowWriter(root, rowType);
         ArrowFieldWriter<RowData>[] fieldWriters = writer.getFieldWriters();
         for (int i = 0; i < fieldWriters.length; i++) {
-            assertEquals(testFields.get(i).f4, fieldWriters[i].getClass());
+            assertEquals(testFields.get(i).f3, fieldWriters[i].getClass());
         }
     }
